커밋해야지

+++ API/demo/src/main/java/com/example/demo/service/FileServiceImpl.java
@@ -52,7 +52,25 @@ public class FileServiceImpl implements FileService {
 
     @Override
     public int delete(int no) {
-        return fileMapper.delete(no);
+        // 1. 파일 정보 조회
+        Files file = fileMapper.select(no);
+        // 2. 파일 경로로 파일 객체 접근
+        String filePath = file.getFilePath();
+        File deleteFile = new File(filePath);
+        // 3. 파일시스템의 파일 삭제
+        // - 파일 존재여부 확인
+        if(!deleteFile.exists()) return 0;
+        // - 파일삭제
+        boolean deleted = deleteFile.delete();
+
+        int result = 0;
+        // DB의 파일 데이터 삭제
+        if(deleted) {
+            result = fileMapper.delete(no);
+            return result;
+        }
+
+        return result;
     }
 
     @Override
@@ -159,6 +177,25 @@ public class FileServiceImpl implements FileService {
         sos.close();
         return 1;
         
+    }
+
+    @Override
+    public int deleteFiles(String no) throws Exception {
+
+        String[] noList = no.split(",");
+
+        int result = 0;
+        for (String deleteNo : noList) {
+            int fileNo = Integer.parseInt(deleteNo);
+            result += delete(fileNo);            
+        }
+
+        return result;
+    }
+
+    @Override
+    public int deleteByParent(Files file) {
+        return fileMapper.deleteByParent(file);
     }
     
 }
